requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
crewai>=0.11.0
crewai-tools>=0.0.5
pydantic>=2.4.2
python-dotenv>=1.0.0 